Check if `doc_user` table exists before drop
Issue #976

+++ server/database/migrations/2016_05_19_160408_move_individual_sponsor_docs_to_individual_groups.php
@@ -34,7 +34,9 @@ class MoveIndividualSponsorDocsToIndividualGroups extends Migration
             ]);
         }
 
-        Schema::drop('doc_user');
+        if (Schema::hasTable('doc_user')) {
+            Schema::drop('doc_user');
+        }
     }
 
     /**
